requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
python-dotenv
streamlit
folium
streamlit_folium
sklearn.preprocessing